+++ src/globals/index.js
@@ -1,9 +1,10 @@
 import globalConfig from './globalConfig.json';
-import * as Errors from './errors';
 
 const stage = process.env.REACT_APP_USER_POOL || process.env.REACT_APP_STAGE;
+const domain = process.env.REACT_APP_DOMAIN || '.dragonchain.com';
 const env = (stage === 'production' || stage === 'local' || stage === 'development') ? stage : 'local';
 
 const GlobalConfig = globalConfig[env];
 
-export { GlobalConfig, Errors, env };
+
+export { GlobalConfig, domain, env };